# src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  version: 1.0.0
  description: API documentation for the E-commerce platform
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        type:
          type: string
          enum: [OK, ERROR]
        message:
          type: string
        data:
          type: object
        meta:
          type: object
          properties:
            id:
              type: string
              format: uuid
...
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string
          format: uri
        website:
          type: string
          format: uri
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
paths:
...
  /brands/{id}/orders:
    post:
      tags:
        - Brands
      summary: Add an order placed (for a brand)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the brand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
                - items
                - totalAmount
              properties:
                orderId:
                  type: string
                  description: Order ID (required)
                status:
                  type: string
                  enum: [pending, processing, completed, cancelled]
                  default: pending
                  description: Order status (optional)
                deliveryDate:
                  type: string
                  format: date-time
                  description: Delivery date (optional)
                notes:
                  type: string
                  description: Order notes (optional)
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                      - price
                    properties:
                      productId:
                        type: string
                        description: Product ID (required)
                      quantity:
                        type: number
                        minimum: 1
                        description: Quantity (required)
                      price:
                        type: number
                        minimum: 0
                        description: Price (required)
                totalAmount:
                  type: number
                  minimum: 0
                  description: Total order amount (required)
      responses:
        '201':
          description: Order added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
              $ref: '#/components/schemas/Brand'
        '404':
          description: Brand not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error (e.g. "Failed to add order")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /brands/{brandId}/orders/{orderId}:
    put:
      tags:
        - Brands
      summary: Update an order placed's status (for a brand)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: brandId
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the brand
        - in: path
          name: orderId
          required: true
          schema:
            type: string
          description: ID of the order (as stored in the brand's orderPlaced array)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, processing, completed, cancelled]
                  description: New status (required)
                deliveryDate:
                  type: string
                  format: date-time
                  description: Delivery date (optional)
                notes:
                  type: string
                  description: Order notes (optional)
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
                properties:
                  data:
                    $ref: '#/components/schemas/Brand'
        '404':
          description: Brand or order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: Internal server error (e.g. "Failed to update order status")
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...